Rename and relocate ID resolution helpers
